Fragment tabs
Tabs work. Need to add details on how to use and on settings menu
