+++ app/src/main/java/com/example/philip/mygpsapp/MainFragment.java
@@ -1,6 +1,6 @@
 package com.example.philip.mygpsapp;
 
-import android.app.Fragment;
+import android.support.v4.app.Fragment;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,6 +10,11 @@ import android.view.ViewGroup;
  * Created by Phil on 11/11/2015.
  */
 public class MainFragment extends Fragment{
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
     @Override
     public View onCreateView(LayoutInflater inflater,
                              ViewGroup container,
