+++ app/src/main/res/layout/activity_main.xml
@@ -1,70 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity">
+<android.support.v4.app.FragmentTabHost
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@android:id/tabhost"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
 
-    <Button
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:text="Main menu"
-        android:id="@+id/btnMain"
-        android:onClick="selectFragment"/>
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="View Settings"
-        android:id="@+id/btnSettings"
-        android:onClick="selectFragment"
-        android:layout_below="@+id/btnCancel"
-        android:layout_alignParentStart="true" />
-
-    <Button
-        style="?android:attr/buttonStyleSmall"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Cancel data gathering"
-        android:id="@+id/btnCancel"
-        android:layout_below="@+id/btnGatherData"
-        android:layout_alignEnd="@+id/btnGatherData" />
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/connectToDrone"
-        android:id="@+id/btnConnect"
-        android:layout_above="@+id/btnSettings"
-        android:layout_alignParentStart="true" />
-
-    <Button
-        android:id="@+id/btnGatherData"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:text="Start gathering data"
-        android:layout_below="@+id/btnMain"
-        android:layout_alignParentStart="true" />
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="View Data"
-        android:id="@+id/btnDataDetails"
-        android:onClick="selectFragment"
-        android:layout_above="@+id/fragment_place"
-        android:layout_centerHorizontal="true" />
-
-    <fragment
-        android:name="com.example.philip.mygpsapp.MainFragment"
-        android:id="@+id/fragment_place"
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_below="@+id/btnSettings" />
+        android:orientation="vertical">
 
+        <TabWidget
+            android:id="@android:id/tabs"
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"/>
 
+        <FrameLayout
+            android:id="@android:id/tabcontent"
+            android:layout_width="fill_parent"
+            android:layout_height="0dp"
+            android:layout_weight="1"/>
+    </LinearLayout>
 
-</RelativeLayout>
+</android.support.v4.app.FragmentTabHost>
