+++ app/src/main/res/layout/fragment_settings.xml
@@ -5,11 +5,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Settings"
         android:id="@+id/textView12"
         android:layout_gravity="center_horizontal" />
+
 </LinearLayout>
